% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getTimeseries}
\alias{getTimeseries}
\title{Get time series of summary functions from simulation}
\usage{
getTimeseries(sim, func = getRDI, ...)
}
\arguments{
\item{sim}{A MizerSim object}

\item{func}{The function calculating the quantities at a single time step}

\item{...}{Unused}
}
\value{
A matrix (time x species)
}
\description{
Given a function that can calculate a vector of quantities of interest at a
single time for all species, this function returns an array (time x species)
with the values at all the time steps saved in a simulation.
}
