% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{gonadDynamics}
\alias{gonadDynamics}
\title{Update the gonadic mass using a seasonal reproduction rate}
\usage{
gonadDynamics(params, n_other, rates, t, dt, ...)
}
\arguments{
\item{params}{MizerParams object}

\item{n_other}{Other model components}

\item{rates}{Previously calculated rates, including in particular e_repro}

\item{t}{The current time}

\item{dt}{The time step size}

\item{...}{Unused}
}
\value{
Array (species x size) with the current gonadic mass of an individual.
}
\description{
Update the gonadic mass using a seasonal reproduction rate
}
