% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsed.R
\name{gonadPulsedDynamics}
\alias{gonadPulsedDynamics}
\title{Update the gonadic mass to the next time step}
\usage{
gonadPulsedDynamics(params, n_other, rates, t, dt, ...)
}
\arguments{
\item{params}{MizerParams object}

\item{n_other}{Other model components}

\item{rates}{Previously calculated rates, including in particular e_repro}

\item{t}{The current time}

\item{dt}{The time step size}

\item{...}{Unused}
}
\value{
Array (species x size) with the current gonadic mass of an individual.
}
\description{
If it is time for the reproductive pulse, the gonadic mass is set to zero,
otherwise it is increased by the energy available for reproduction.
}
