% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alt_seasonal_functions.R
\name{plotGonads}
\alias{plotGonads}
\title{not a real function. Used to follow gonadic mass through time}
\usage{
plotGonads(
  sim,
  time_range,
  prop_size = 1,
  sizes = sim@params@species_params$w_max,
  ...
)
}
\description{
not a real function. Used to follow gonadic mass through time
}
