% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsed.R
\name{pulsedRDI}
\alias{pulsedRDI}
\title{Get density-dependent rate of pulsed reproduction}
\usage{
pulsedRDI(params, n_other, t, dt, ...)
}
\arguments{
\item{params}{MizerParams object}

\item{n_other}{Other model components}

\item{t}{The current time}

\item{dt}{The time step size}

\item{...}{Unused}
}
\value{
A numeric vector with the rate of egg production for each species.
}
\description{
The rate of reproduction is zero except during one timestep, during which
all gonadic mass is released.
}
