% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alt_seasonal_functions.R
\name{repro_beta_hazard}
\alias{repro_beta_hazard}
\title{Beta hazard mass-specific reproduction rate}
\usage{
repro_beta_hazard(t, params)
}
\arguments{
\item{t}{The time at which to calculate the reproduction rate}

\item{params}{A MizerParams object}
}
\value{
A vector of species-specific reproduction rates
}
\description{
Uses the formula
}
