% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{repro_gaussian}
\alias{repro_gaussian}
\title{Gaussian mass-specific reproduction rate}
\usage{
repro_gaussian(t, params)
}
\arguments{
\item{t}{The time at which to calculate the reproduction rate}

\item{params}{A MizerParams object}
}
\value{
A vector of species-specific reproduction rates
}
\description{
\deqn{r(w, t) = r_0 \exp{\left(-\frac{(\floor{t}-t_0)^2}{2\sigma^2}\right)}}{r_0 \exp(-(\floor{t}-t_0)^2/(2\sigma^2))}
}
