% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alt_seasonal_functions.R
\name{repro_vonMises}
\alias{repro_vonMises}
\title{von-Mises distributed reproduction rate}
\usage{
repro_vonMises(t, params, ...)
}
\arguments{
\item{t}{The time at which to calculate the reproduction rate}

\item{params}{A MizerParams object}
}
\value{
A vector of species-specific reproduction rates
}
\description{
Uses the formula
}
