% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonal_resource_semichemostat}
\alias{seasonal_resource_semichemostat}
\title{Seasonal semichemostat resource dynamics}
\usage{
seasonal_resource_semichemostat(
  params,
  n,
  n_pp,
  n_other,
  rates,
  t,
  dt,
  resource_rate,
  resource_capacity,
  ...
)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list with the abundances of other components}

\item{rates}{A list of rates as returned by \code{\link[=mizerRates]{mizerRates()}}}

\item{t}{The current time}

\item{dt}{Time step}

\item{resource_rate}{Resource replenishment rate}

\item{resource_capacity}{Resource carrying capacity}

\item{...}{Unused}
}
\value{
Vector containing resource spectrum at next timestep
}
\description{
This implements the standard semichemostat dynamics for the resource (see
\code{mizer::resource_semichemostat()}) but with a time-dependent carrying
capacity. The carrying capacity is given by \deqn{c_R(w,t) = (1 + \text{maxR}
\cdot \text{vonMises}(t, \mu, \kappa)) c_R(w)}{c_R(w,t) = (1 + maxR * vonMises(t,
mu, kappa)) c_R(w)} where \eqn{c_R(w)} is the standard carrying capacity,
\eqn{\mu}{mu} is the mean of the von-Mises distribution, and
\eqn{\kappa}{kappa} is the concentration parameter of the von-Mises
distribution. These parameters must be supplied in the \code{rp$maxR}, \code{rp$mu} and
\code{rp$kappa} slots of the \code{resource_params} slot of the \code{params} object.
}
\concept{resource dynamics}
