% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alt_seasonal_functions.R
\name{seasonalBetaRelease}
\alias{seasonalBetaRelease}
\title{Beta distributed release rate}
\usage{
seasonalBetaRelease(t, params)
}
\arguments{
\item{t}{The time at which to calculate the reproduction rate}

\item{params}{A MizerParams object}
}
\value{
A vector of species-specific release rates at time t
}
\description{
Beta distributed release rate
}
