% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonalEncounter}
\alias{seasonalEncounter}
\title{Get encounter rate that includes gonadic mass of prey}
\usage{
seasonalEncounter(params, n, n_pp, n_other, t, ...)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation}

\item{...}{Unused}
}
\value{
A named two dimensional array (predator species x predator size) with
the encounter rates.
}
\description{
This is doing the same as the \code{mizerEncounter()} function in core mizer
except that the prey mass is the sum of its somatic mass \eqn{w_p} and its
gonadic mass \eqn{q(w_p)}. This function is automatically registered as the
model's encounter rate function when \code{setSeasonalReproduction()} is called
with \code{include_gonads = TRUE}.
}
\concept{mizer rate functions}
