% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonalRates}
\alias{seasonalRates}
\title{Seasonal version of mizerRates()}
\usage{
seasonalRates(params, n, n_pp, n_other, t = 0, effort, rates_fns, ...)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{effort}{The effort for each fishing gear}

\item{rates_fns}{Named list of the functions to call to calculate the rates.
Note that this list holds the functions themselves, not their names.}

\item{...}{Unused}
}
\value{
List of rates.
}
\description{
This is needed because we need to pass the time argument to the RDD function
}
