% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonalRDI}
\alias{seasonalRDI}
\title{Get density-independent rate of seasonal reproduction}
\usage{
seasonalRDI(params, n, n_other, t, dt = 0.1, ...)
}
\arguments{
\item{params}{MizerParams object}

\item{n}{Species abundances at current time step}

\item{n_other}{Other model components at current time step. This will include
in particular the gonadic biomass in \code{n_other$gonads}.}

\item{t}{The current time}

\item{dt}{The time step size}

\item{...}{Unused}
}
\value{
A numeric vector with the rate of egg production for each species.
}
\description{
This function calculates the density-independent rate of offspring production
for each species. The rate of offspring production is given by
\deqn{R_{di}(t) =\frac{\epsilon}{2w_0}\int N(w,t)q(w,t)r(w,t)dw}{R_di(t) =
\epsilon/(2w_0) int N(w,t)gw(w,t)r(w,t)dw}
where \eqn{q(w,t)} is the gonadic mass of an individual, \eqn{r(w,t)}{r(w,t)}
is the mass-specific gonad release rate, \eqn{N(w,t)}{N(w,t)} is the
abundance density, \eqn{\epsilon}{epsilon} is the reproductive efficiency and
\eqn{w_0}{w_0} is the weight of an offspring.
}
