% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonalRepro}
\alias{seasonalRepro}
\title{Seasonal reproduction rate}
\usage{
seasonalRepro(t, params)
}
\arguments{
\item{t}{The time at which to calculate the reproduction rate}

\item{params}{A MizerParams object}
}
\value{
A vector of species-specific reproduction rates
}
\description{
Calculate the mass-specific reproduction rate as a function of time
}
