% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonalVonMisesRDD}
\alias{seasonalVonMisesRDD}
\title{von-Mises distributed reproduction rate independent of abundance}
\usage{
seasonalVonMisesRDD(params = NULL, t = NULL, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{t}{The time at which to calculate RDD}
}
\value{
A vector of species-specific reproduction rates
}
\description{
von-Mises distributed reproduction rate independent of abundance
}
