% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_functions.R
\name{seasonalVonMisesRelease}
\alias{seasonalVonMisesRelease}
\title{von-Mises distributed gonad release rate}
\usage{
seasonalVonMisesRelease(t, params, ...)
}
\arguments{
\item{t}{The time at which to calculate the release rate}

\item{params}{A MizerParams object}

\item{...}{Unused}
}
\value{
A vector of species-specific release rates at time t
}
\description{
Uses the formula
\deqn{r(w, t) = r(t) = r_0 \frac{\exp(\kappa \cos(2\pi(t - \mu)))}{2\pi I_0(\kappa)}}{r(w, t) = r(t) = r_0 exp(\kappa cos(2\pi(t - \mu))) / (2\pi I_0(\kappa))}
where the parameters are taken from the \code{vonMises_r0}, \code{vonMises_kappa} and
\code{vonMises_mu} columns in the \code{species_params} data frame in \code{params}.
}
\seealso{
Other release functions: 
\code{\link{seasonalBetaHazardRelease}()},
\code{\link{seasonalBetaRelease}()},
\code{\link{seasonalGaussianRelease}()}
}
\concept{release functions}
