% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsed.R
\name{setPulsedReproduction}
\alias{setPulsedReproduction}
\title{Set pulsed reproduction}
\usage{
setPulsedReproduction(params, pulse_time)
}
\arguments{
\item{params}{A MizerParams object}

\item{pulse_time}{A number between 0 and 1 that specifies the time within
each year at which reproduction should take place.}
}
\value{
A MizerParams object with the updated model
}
\description{
This returns a new model in which the reproduction is set to take place
only within a single time step each year. During the rest of the year the energy
available for reproduction is accumulated as gonadic mass, all of which is
realeased during the time step into which the \code{pulse_time} falls.
}
