% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{setSeasonalReproduction}
\alias{setSeasonalReproduction}
\title{Set seasonal reproduction}
\usage{
setSeasonalReproduction(params, pulse_time)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A MizerParams object with seasonal reproduction
}
\description{
This returns a new model in which the reproduction rate varies throughout
the year
}
