% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMises.R
\name{setVonMisesReproduction}
\alias{setVonMisesReproduction}
\title{Set reproduction with vonMises distribution}
\usage{
setVonMisesReproduction(params, pulse_time)
}
\arguments{
\item{params}{A MizerParams object}

\item{pulse_time}{A number between 0 and 1 that specifies the time within
each year at which reproduction should take place.}
}
\value{
A MizerParams object with von Mises reproduction rate
}
\description{
This returns a new model in which the reproduction rate varies throughout
the year so that
}
