% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datta_params}
\alias{datta_params}
\title{Parameters used by Datta & Blanchard (2016)}
\format{
A MizerParams object
}
\source{
Datta, S. & Blanchard, J. L. "The effects of seasonal processes on
size spectrum dynamics". Canadian Journal of Fisheries and Aquatic Sciences
(2016). \url{https://cdnsciencepub.com/doi/full/10.1139/cjfas-2015-0468}
}
\usage{
datta_params
}
\description{
This MizerParams object is an approximation of the model used by Datta &
Blanchard (2016), which allows to reproduce their results and to extend it
to the full gonadic mass dynamics implemented in mizerSeasonal.
}
\details{
The paper "The effects of seasonal processes on size spectrum dynamics" by
Samik Datta and Julia Blanchard, published in Can. J. Fish. Aquat. Sci. 73,
598–610 (2016), was the first paper to investigate seasonal processes in
the context of size spectrum dynamics. The authors used an early version of
the mizer code that they then modified. Their code is available at
\url{https://figshare.com/s/e75f29d4cc9b94ae393b}.

This MizerParams object is based on the parameters used in the paper.
However the choice of size bins for the numerical implementation of the
mizer equations is different from the one used in the paper. The paper used
wider size-bins for the resource spectrum than for the fish spectrum. The
modern mizer code no longer supports this. So this MizerParams object uses
the same size bins for the resource and fish spectra. Like the model by
Datta and Blanchard, we use 100 size bins for the fish spectra but we use
180 size bins for the full spectra instead of 130.
The initial state of the MizerParams object is set to the steady state of the
model.

The script that created this MizerParams object is available at
\url{https://github.com/gustavdelius/mizerSeasonal/blob/main/data-raw/datta_params.R}.
}
\keyword{datasets}
