% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datta_params_second_order}
\alias{datta_params_second_order}
\title{Second-order parameters used by Datta & Blanchard (2016)}
\format{
A \link[mizer:MizerParams]{mizer::MizerParams} object, stored as an S3 object like
\link{datta_params}.
}
\source{
Datta, S. & Blanchard, J. L. "The effects of seasonal processes on
size spectrum dynamics". Canadian Journal of Fisheries and Aquatic Sciences
(2016). \url{https://cdnsciencepub.com/doi/full/10.1139/cjfas-2015-0468}
}
\usage{
datta_params_second_order
}
\description{
A variant of \link{datta_params} created with \code{second_order_w = TRUE}. It uses
the second-order van Leer flux scheme and bin-averaged quantities, and its
initial state is set to the steady state of the second-order model.
}
\keyword{datasets}
