% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerSeasonal-class.R
\name{mizerSeasonal-class}
\alias{mizerSeasonal-class}
\title{mizerSeasonal marker classes}
\description{
S4 marker subclasses of \code{MizerParams} and \code{MizerSim} that enable S3 dispatch
for this package's methods. They add no slots and are created dynamically by
mizer when the package is loaded.
}
\keyword{internal}
