% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerSeasonal-package.R
\docType{package}
\name{mizerSeasonal-package}
\alias{mizerSeasonal}
\alias{mizerSeasonal-package}
\title{mizerSeasonal: Extends the mizer package to allow for seasonal reproduction}
\description{
This is an extension package for the
\href{https://sizespectrum.org/mizer/}{mizer package}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gustavdelius/mizerSeasonal}
  \item Report bugs at \url{https://github.com/sizespectrum/mizerSeasonal/issues}
}

}
\author{
\strong{Maintainer}: Gustav Delius \email{gustav.delius@york.ac.uk} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Michael A. Spence \email{michael.spence@cefas.gov.uk} (\href{https://orcid.org/0000-0002-3445-7979}{ORCID}) [copyright holder]
  \item Caroline McKeon \email{mckeonc2@tcd.ie} (\href{https://orcid.org/0000-0002-1737-3503}{ORCID}) [copyright holder]
}

}
\keyword{internal}
