% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{projectEncounter.mizerSeasonal}
\alias{projectEncounter.mizerSeasonal}
\title{Seasonal encounter rate including gonadic mass of prey via S3 dispatch}
\usage{
\method{projectEncounter}{mizerSeasonal}(params, n, n_pp, n_other, t = 0, ...)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object of class \code{mizerSeasonal}.}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size.}

\item{n_other}{A list of other model components, including \code{gonads}.}

\item{t}{The current time. Unused.}

\item{...}{Passed to \code{NextMethod()}.}
}
\value{
A named two-dimensional array (predator species x predator size)
with the encounter rates.
}
\description{
Method for \code{projectEncounter()} that adds the extra encounter arising from
the gonadic mass carried by prey fish. It first calls \code{NextMethod()} to
obtain the standard encounter (using somatic mass only), then appends the
contribution from the gonadic mass of prey fish. This is only done when
\code{include_gonads = TRUE} was passed to \code{\link[=setSeasonalReproduction]{setSeasonalReproduction()}}.
}
