% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonal_resource_semichemostat}
\alias{seasonal_resource_semichemostat}
\title{Seasonal semichemostat resource dynamics}
\usage{
seasonal_resource_semichemostat(
  params,
  n,
  n_pp,
  rates,
  t,
  dt,
  resource_rate,
  resource_capacity,
  ...
)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of species abundances (species x size)}

\item{n_pp}{A vector of the resource abundance by size}

\item{rates}{A list of rates as returned by \code{\link[=mizerRates]{mizerRates()}}}

\item{t}{The current time}

\item{dt}{Time step}

\item{resource_rate}{Resource replenishment rate}

\item{resource_capacity}{Resource carrying capacity}

\item{...}{Unused}

\item{n_other}{A list with the abundances of other components}
}
\value{
Vector containing resource spectrum at next timestep
}
\description{
Seasonal semichemostat resource dynamics
}
\concept{resource dynamics}
