% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_functions.R
\name{seasonalBetaHazardRelease}
\alias{seasonalBetaHazardRelease}
\title{Beta hazard mass-specific gonad release rate}
\usage{
seasonalBetaHazardRelease(t, params)
}
\arguments{
\item{t}{The time at which to calculate the reproduction rate}

\item{params}{A MizerParams object}
}
\value{
A vector of species-specific release rates at time t
}
\description{
Beta hazard mass-specific gonad release rate
}
