% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_functions.R
\name{seasonalBetaRelease}
\alias{seasonalBetaRelease}
\title{Beta distributed gonad release rate}
\usage{
seasonalBetaRelease(t, params)
}
\arguments{
\item{t}{The time at which to calculate the reproduction rate}

\item{params}{A MizerParams object}
}
\value{
A vector of species-specific release rates at time t
}
\description{
Uses the formula
\deqn{r(w, t) = r(t) = f(t-\lfloor t \rfloor)}{r(w,t) = r(t) = f(t-|_t_|)}
where \eqn{f(t)} is the probability density function of the beta distribution
(see see \code{?dbeta}) with parameters \code{shape1 = beta_a} and \code{shape2 = beta_b}.
Because mizer measures time in years, \eqn{t-\lfloor t \rfloor}{t-|_t_|}
gives the time within the year and so \eqn{r(t)} is a periodic function with
the period of one year.
}
\seealso{
Other release functions: 
\code{\link{seasonalBetaHazardRelease}()},
\code{\link{seasonalGaussianRelease}()},
\code{\link{seasonalVonMisesRelease}()}
}
\concept{release functions}
