% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonalRDI}
\alias{seasonalRDI}
\title{Get density-independent rate of seasonal reproduction}
\usage{
seasonalRDI(params, n, n_other, t, dt = 0.1, ...)
}
\arguments{
\item{params}{MizerParams object}

\item{n}{Species abundances at current time step}

\item{n_other}{Other model components at current time step. This will include
in particular the gonadic biomass in \code{n_other$gonads}.}

\item{t}{The current time}

\item{dt}{The time step size}

\item{...}{Unused}
}
\value{
A numeric vector with the rate of egg production for each species.
}
\description{
Get density-independent rate of seasonal reproduction
}
