% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonalRepro}
\alias{seasonalRepro}
\title{Seasonal reproduction rate}
\usage{
seasonalRepro(params, t)
}
\arguments{
\item{params}{A MizerParams object}

\item{t}{The time at which to calculate the reproduction rate}
}
\value{
A vector of species-specific reproduction rates
}
\description{
Calculate the mass-specific reproduction rate as a function of time
}
