% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{setSeasonalReproduction}
\alias{setSeasonalReproduction}
\title{Set seasonal reproduction}
\usage{
setSeasonalReproduction(params, repro_func = "repro_gaussian")
}
\arguments{
\item{params}{A MizerParams object}

\item{repro_func}{Name of the function giving the time-dependent
mass-specific reproduction rate.}
}
\value{
A MizerParams object with seasonal reproduction
}
\description{
This returns a new model in which the reproduction rate varies throughout
the year
}
