% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{setSeasonalReproduction}
\alias{setSeasonalReproduction}
\title{Set seasonal reproduction}
\usage{
setSeasonalReproduction(
  params,
  release_func = "repro_vonMises",
  RDD = "seasonalBevertonHoltRDD",
  include_gonads = TRUE
)
}
\arguments{
\item{params}{A MizerParams object}

\item{release_func}{Name of the function giving the time-dependent
mass-specific reproduction rate.}

\item{RDD}{Name of the function for calculating the density-dependent
reproduction rate RDD.}

\item{include_gonads}{Boolean. If TRUE (default) then the gonadic mass is
included in the prey encounter rate.}
}
\value{
A MizerParams object with seasonal reproduction
}
\description{
This returns a new model in which the reproduction rate varies throughout
the year
}
